LayoutTests/fast/events/touch/ios/pencil-data-change-gives-touchforcechange-expected.txt

Test that an Apple Pencil produces touchforcechange events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



    touchstart fired.
PASS event.touches.length is 1
PASS event.touches[0].touchType is 'stylus'
PASS event.touches[0].altitudeAngle is within 0.01 of 1
PASS event.touches[0].azimuthAngle is within 0.01 of 2
PASS event.touches[0].force is within 0.01 of 0.5
touchforcechange fired.
PASS event.touches.length is 1
PASS event.touches[0].touchType is 'stylus'
PASS event.touches[0].altitudeAngle is within 0.01 of 1.1
PASS event.touches[0].azimuthAngle is within 0.01 of 2.1
PASS event.touches[0].force is within 0.01 of 0.6
touchend fired.
PASS successfullyParsed is true

TEST COMPLETE
